backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Async file operations
aiofiles>=24.0.0

# Database
sqlalchemy>=2.0.43
asyncpg>=0.30.0
alembic>=1.16.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Pydantic and Settings
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Environment and Configuration
python-dotenv>=1.0.1

# AWS and Cloud Services
boto3>=1.35.0
botocore>=1.35.0

# HTTP Client
httpx>=0.27.0
requests>=2.32.0
requests-oauthlib>=2.0.0

# Web scraping and HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Development and Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
black>=24.4.0
isort>=5.13.0
flake8>=7.1.0
mypy>=1.11.0
faker>=26.0.0

# AI and ML
openai>=1.40.0

# Export functionality for test cases
openpyxl>=3.1.5  # Excel export
reportlab>=4.2.0  # PDF export
pandas>=2.2.0  # Data manipulation for exports
fpdf2>=2.7.9  # Alternative PDF library

# Utilities
email-validator>=2.2.0
python-dateutil>=2.9.0
typing-extensions>=4.12.0
starlette>=0.40.0
concurrent-log-handler>=0.9.28

# Temporarily commented out as it's not available on PyPI
# emergentintegrations